pyproject.toml
[project]
name = "laibrary"
version = "0.1.0"
description = "An AI-curated library of your thoughts and ideas"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core
    "langgraph>=1.0.7",
    "pydantic-ai>=1.47.0",
    "logfire>=3.0.0",
    "python-dotenv>=1.0.0",
    # Graph RAG
    "networkx>=3.0",
    "chromadb>=0.4.0",
    "ollama>=0.3.0",
    # CLI
    "typer>=0.12.0",
    "rich>=13.0",
    # Version control
    "gitpython>=3.1",
]

[project.scripts]
laibrary = "laibrary.cli:main"

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
# E/W: Errors/Warnings, F: Pyflakes, I: Isort, B: Bugbear, N: PEP8-Naming, D: Docstrings, UP: Unused imports and variables
select = ["E", "W", "F", "I", "B", "N", "D", "UP"]
ignore = [
    "B008",  # Ignore "function call in default argument" (common in Typer)
    "D202",  # Ignore "No blank lines allowed after function docstring"
    "D203",  # Ignore "1 blank line required before class docstring"
    "D213",  # Ignore "Multi-line docstring summary should start at the second line"
    "E501",  # Ignore line length in all code (docstrings need to be parseable)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.0",
    "pre-commit>=4.5.1",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
